# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mentors:
    get:
      operationId: listMentors
      tags: [mentors]
      summary: List mentors
      parameters:
        - name: skills
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Mentors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MentorProfile"
  /mentors/profile:
    get:
      operationId: getMentorProfile
      tags: [mentors]
      summary: Get current mentor profile
      responses:
        "200":
          description: Mentor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
    put:
      operationId: updateMentorProfile
      tags: [mentors]
      summary: Update mentor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MentorProfileInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"
  /mentors/{mentorId}:
    get:
      operationId: getMentorById
      tags: [mentors]
      summary: Get mentor by ID
      parameters:
        - name: mentorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mentor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MentorProfile"

  # Jobs
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{jobId}:
    get:
      operationId: getJobById
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourseById
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}/enroll:
    post:
      operationId: enrollCourse
      tags: [courses]
      summary: Enroll in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseEnrollment"
  /courses/{courseId}/progress:
    patch:
      operationId: updateCourseProgress
      tags: [courses]
      summary: Update course progress
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseProgressUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseEnrollment"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth
...
    MentorProfile:
      type: object
      required: [id, userId, name, email]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        experience:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        sessionsCompleted:
          type: integer
        pricePerSession:
          type: ["number", "null"]
        availability:
          type: array
          items:
            type: string
    MentorProfileInput:
      type: object
      properties:
        company:
          type: string
        title:
          type: string
        experience:
          type: integer
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        avatarUrl:
          type: string
        pricePerSession:
          type: number
        availability:
          type: array
          items:
            type: string

    # Jobs
    Job:
      type: object
      required: [id, companyId, title, description, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        companyLogoUrl:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        experienceLevel:
          type: ["string", "null"]
        jobType:
          type: ["string", "null"]
        isRemote:
          type: boolean
        isActive:
          type: boolean
        applicantCount:
          type: integer
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        salary:
          type: string
        location:
          type: string
        experienceLevel:
          type: string
        jobType:
          type: string
        isRemote:
          type: boolean
        isActive:
          type: boolean

    # Applications
...
    Course:
      type: object
      required: [id, title, description, category]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        instructor:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        level:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        enrollmentCount:
          type: integer
        skills:
          type: array
          items:
            type: string
        isFree:
          type: boolean
        price:
          type: ["number", "null"]
    CourseEnrollment:
      type: object
      required: [id, courseId, studentId, progress, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        studentId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        courseThumbnailUrl:
          type: ["string", "null"]
        courseCategory:
          type: ["string", "null"]
        progress:
          type: integer
        completed:
          type: boolean
        createdAt:
          type: string
    CourseProgressUpdate:
      type: object
      required: [progress]
      properties:
        progress:
          type: integer
        completed:
          type: boolean

    # AI
...