# lib/api-zod/src/generated/types/courseEnrollment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseEnrollment {
  id: number;
  courseId: number;
  studentId: number;
  /** @nullable */
  courseTitle?: string | null;
  /** @nullable */
  courseThumbnailUrl?: string | null;
  /** @nullable */
  courseCategory?: string | null;
  progress: number;
  completed?: boolean;
  createdAt: string;
}