# lib/api-zod/src/generated/types/courseProgressUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CareerAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseProgressUpdate {
  progress: number;
  completed?: boolean;
}